add confirmation popup

+++ src/app/brand/brand-table/brand-table.component.ts
@@ -9,6 +9,7 @@ import {Brand} from "../../interfaces/entity/brand";
 import {BrandEditorComponent} from "../brand-editor/brand-editor.component";
 import {CountryService} from "../../services/country.service";
 import {BrandCreateComponent} from "../brand-create/brand-create.component";
+import {ConfirmationPopupComponent} from "../../shared/components/confirmation-popup/confirmation-popup.component";
 
 @Component({
   selector: 'app-brand-table',
@@ -101,8 +102,12 @@ export class BrandTableComponent implements AfterViewInit {
   }
 
   public delete(id: string): void {
-    // TODO: add popup
-    this.brandService.remove(id).subscribe(() => this.getBrands())
-  }
+    const dialogRef = this.dialog.open(ConfirmationPopupComponent);
 
+    dialogRef.afterClosed().subscribe(popupResponse => {
+      if (popupResponse) {
+        this.brandService.remove(id).subscribe(() => this.getBrands())
+      }
+    });
+  }
 }
